refactor(oracle): extract wireshark decode helpers module
